Move JPetParamsFactory to ParametersTools
